Entity model API, REST version demo
I started this exploration using fake data to power a small REST API
based on the JSON API spec, as a way to explore how to represent the
relationships between entities.

I'm going to pivot now and instead build small Kibana JS functions to
return related entities. We can construct a REST API from those
functions if we need it.