Removed unnecessary .keyword distinctions in ES queries
.keyword was introduced in the queries when the mappings were improperly
configured. First attempt to change the mappings failed because it had
subobjects: false in the top level of the dynamic assets mapping template.

I moved subobjects: false into the properties -> asset block instead,
so that ECS fields can be nested as usual, but asset fields will
only work as dotted.